+++ sample/Assets/Script/CameraScript.cs
@@ -53,7 +53,7 @@ public class CameraScript : MonoBehaviour {
 				RelayPos	= transform.position;
 				workPos		= transform.position;
 								
-				transform.position = SLineManager.CalcPos(SPlayerScript.m_Timer-60, 0);
+				transform.position = SLineManager.CalcPos(((int)SPlayerScript.m_Timer)-60, 0);
 				transform.position = new Vector3(	transform.position.x,
 													transform.position.y+3.0f,
 													transform.position.z );
@@ -91,11 +91,11 @@ public class CameraScript : MonoBehaviour {
 				shake.x += offset;
 			}
 			// 刃の座標を求める
-			TargetPos = SLineManager.CalcPos(SPlayerScript.m_Timer, shake.x);
+			TargetPos = SLineManager.CalcPos((int)SPlayerScript.m_Timer, shake.x);
 			TargetPos = new Vector3(TargetPos.x, TargetPos.y+shake.y, TargetPos.z);
 				
 			// 現在の刃の座標から数秒前の位置にカメラをセットする
-			transform.position = SLineManager.CalcPos(SPlayerScript.m_Timer-60, shake.x);
+			transform.position = SLineManager.CalcPos(((int)SPlayerScript.m_Timer-60), shake.x);
 			transform.position = new Vector3(	transform.position.x,
 												transform.position.y+3.0f+shake.y,
 												transform.position.z );
